Add mvc like
