+++ model/people.js
@@ -1,0 +1,5 @@
+var people = require('../people.json')
+
+module.exports = function () {
+    return people
+}
